[Rename] WORKSPACE_005, WORKSPACE_009 - 프로젝트 구조 변경으로 파일 이동
