+++ backend/src/main/java/minionz/backend/scrum/label/model/SprintLabel.java
@@ -1,33 +1,33 @@
-package minionz.backend.scrum.task_label.model;
+package minionz.backend.scrum.label.model;
 
 import jakarta.persistence.*;
 import lombok.*;
-import minionz.backend.scrum.task_label_select.model.TaskLabelSelect;
+import minionz.backend.scrum.sprint_label_select.model.SprintLabelSelect;
 import minionz.backend.scrum.workspace.model.Workspace;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Entity
-public class TaskLabel {
+public class SprintLabel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long taskLabelId;
+    private Long sprintLabelId;
 
-    private String taskName;
+    private String labelName;
     private String description;
     private String color;
 
-    // TaskLabel : TaskLabelSelect = 1 : N
-    @OneToMany(mappedBy = "taskLabel", fetch = FetchType.LAZY)
-    private List<TaskLabelSelect> taskLabelSelects = new ArrayList<>();
+    // SprintLabel : SprintLabelSelect = 1 : N
+    @OneToMany(mappedBy = "sprintLabel", fetch = FetchType.LAZY)
+    private List<SprintLabelSelect> sprintLabelSelects = new ArrayList<>();
 
-    // TaskLabel : Workspace = N : 1
+    // SprintLabel : Workspace = N : 1
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "workspace_id")
     private Workspace workspace;
