+++ backend/src/main/java/minionz/backend/scrum/label/model/TaskLabel.java
@@ -1,8 +1,8 @@
-package minionz.backend.scrum.sprint_label.model;
+package minionz.backend.scrum.label.model;
 
 import jakarta.persistence.*;
 import lombok.*;
-import minionz.backend.scrum.sprint_label_select.model.SprintLabelSelect;
+import minionz.backend.scrum.task_label_select.model.TaskLabelSelect;
 import minionz.backend.scrum.workspace.model.Workspace;
 
 import java.util.ArrayList;
@@ -14,20 +14,20 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class SprintLabel {
+public class TaskLabel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long sprintLabelId;
+    private Long taskLabelId;
 
     private String labelName;
     private String description;
     private String color;
 
-    // SprintLabel : SprintLabelSelect = 1 : N
-    @OneToMany(mappedBy = "sprintLabel", fetch = FetchType.LAZY)
-    private List<SprintLabelSelect> sprintLabelSelects = new ArrayList<>();
+    // TaskLabel : TaskLabelSelect = 1 : N
+    @OneToMany(mappedBy = "taskLabel", fetch = FetchType.LAZY)
+    private List<TaskLabelSelect> taskLabelSelects = new ArrayList<>();
 
-    // SprintLabel : Workspace = N : 1
+    // TaskLabel : Workspace = N : 1
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "workspace_id")
     private Workspace workspace;
