+++ backend/src/main/java/minionz/backend/scrum/sprint_label_select/model/SprintLabelSelect.java
@@ -3,7 +3,7 @@ package minionz.backend.scrum.sprint_label_select.model;
 import jakarta.persistence.*;
 import lombok.*;
 import minionz.backend.scrum.sprint.model.Sprint;
-import minionz.backend.scrum.sprint_label.model.SprintLabel;
+import minionz.backend.scrum.label.model.SprintLabel;
 
 @Getter
 @Setter
