+++ backend/src/main/java/minionz/backend/scrum/task_label_select/model/TaskLabelSelect.java
@@ -3,7 +3,7 @@ package minionz.backend.scrum.task_label_select.model;
 import jakarta.persistence.*;
 import lombok.*;
 import minionz.backend.scrum.task.model.Task;
-import minionz.backend.scrum.task_label.model.TaskLabel;
+import minionz.backend.scrum.label.model.TaskLabel;
 
 @Getter
 @Setter
